refactor: move some methods/types from sync.rs to plumbing.rs, operations.rs
