+++ ic-asset/src/lib.rs
@@ -2,7 +2,9 @@ mod asset_canister;
 mod content;
 mod content_encoder;
 mod convenience;
+mod operations;
 mod params;
+mod plumbing;
 mod sync;
 
 pub use sync::sync;
